feat: implement ProjectCard and ProjectModal components with project details and modal functionality
