+++ frontend/src/pages/Work.tsx
@@ -1,24 +1,14 @@
 import { useState } from 'react';
-import { motion } from 'framer-motion';
 import { ArrowUpRight, X } from 'lucide-react';
 
 import projects from '../lib/projects.json';
 import { Project } from '../lib/types';
 
+import ProjectModal from '@/components/ProjectModal';
 import ProjectCard from '@/components/ProjectCard';
+
 import { Button } from '@/components/ui/button';
 
-const itemVariants = {
-  hidden: { opacity: 0, x: 50 },
-  visible: {
-    opacity: 1,
-    x: 0,
-    transition: {
-      duration: 0.6,
-      ease: [0.42, 0, 0.58, 1] as [number, number, number, number],
-    },
-  },
-};
 
 const Work = () => {
   const [selectedProject, setSelectedProject] = useState<Project | null>(null);
@@ -64,28 +54,11 @@ const Work = () => {
 
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 ">
               {projects.slice(3, projects.length).map((project: Project) => (
-                <div
+                <ProjectCard
                   key={project.title}
-                  className="flex flex-col gap-2 justify-between items-center p-4 max-w-sm rounded overflow-hidden shadow-lg hover:shadow-xl transition-shadow duration-600 ease-in-out cursor-pointer bg-white border-primaryColor border outline-2 outline-offset-4"
-                >
-                  <img
-                    src={project.img}
-                    alt={project.title}
-                    className="w-full"
-                  />
-                  <div className="mx-2">
-                    <div className="font-bold text-xl mb-2">{project.title}</div>
-                    <p className="text-gray-700 text-sm ">
-                      {project.desc.split(' ').slice(0, 20).join(' ')}...
-                    </p>
-                  </div>
-                  <div className='button-container flex flex-col items-center gap-4 transition-all duration-300 ease-in-out  bg-black w-full text-white rounded py-1 group'>
-                    <button className='button flex items-center justify-between gap-1' onClick={() => handleButtonClick(project)}>
-                      <p className='text-sm'>Learn More</p>
-                      <ArrowUpRight className='w-4 h-4 transition-transform duration-300 ease-in-out group-hover:rotate-45 ' />
-                    </button>
-                  </div>
-                </div>
+                  project={project}
+                  onClick={() => handleButtonClick(project)}
+                />
               ))}
             </div>
 
@@ -95,35 +68,11 @@ const Work = () => {
 
       <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8 px-8">
         {projects.slice(0, 3).map((project: Project) => (
-          <motion.div
+          <ProjectCard
             key={project.title}
-            className="flex flex-col gap-2 justify-between items-center p-4 max-w-sm rounded overflow-hidden shadow-lg hover:shadow-xl transition-shadow duration-600 ease-in-out cursor-pointer bg-white border-primaryColor border outline-2 outline-offset-4"
-            variants={itemVariants}
-            initial='hidden'
-            whileInView='visible'
-            viewport={{ once: true, amount: 0.2 }}
-          >
-            <img
-              src={project.img}
-              alt={project.title}
-              className="w-full"
-            />
-
-            <div className="mx-2">
-              <div className="font-bold text-xl mb-2">{project.title}</div>
-              <p className="text-gray-700 text-sm ">
-                {project.desc.split(' ').slice(0, 20).join(' ')}...
-              </p>
-            </div>
-
-            <div className='button-container flex flex-col items-center gap-4 transition-all duration-300 ease-in-out  bg-black w-full text-white rounded py-1 group'>
-              <button className='button flex items-center justify-between gap-1' onClick={() => handleButtonClick(project)}>
-                <p className='text-sm'>Learn More</p>
-                <ArrowUpRight className='w-4 h-4 transition-transform duration-300 ease-in-out group-hover:rotate-45 ' />
-              </button>
-            </div>
-
-          </motion.div>
+            project={project}
+            onClick={() => handleButtonClick(project)}
+          />
         ))}
       </div>
     </>
@@ -133,16 +82,24 @@ const Work = () => {
     <section id='work'
       className={'transition-opacity duration-1000 ease-in-out mx-auto pb-8 '}
     >
-      <div className='px-4 sm:px-0'>
-        <p className="section-header sm:ml-8 mt-8">Projects</p>
-        <p className='sm:ml-8 mb-8 text-sm'>
-          Here are some of the projects I have worked on. Click on the "Learn More" button to see more details about each project.
-          <br />
-        </p>
+      <div className='flex justify-around items-center px-8'>
+
+        <div className='px-4 sm:px-0 mx-auto w-full'>
+          <p className="section-header mt-8">Projects</p>
+          <p className=' mb-8 text-sm'>
+            Here are some of the projects I have worked on. Click on the "Learn More" button to see more details about each project.
+            <br />
+          </p>
+        </div>
+        <div className='justify-center mt-8 hidden sm:flex'>
+          <Button variant="link" className='text-primaryColor  border border-primaryColor rounded-xl bg-secondaryColor group' onClick={() => handleSeeMore(projects)}>Click to see more
+            <ArrowUpRight className='w-4 h-4 inline-block transition-transform duration-300 ease-in-out group-hover:rotate-45' />
+          </Button>
+        </div>
       </div>
       {projectList}
 
-      <div className='flex justify-center mt-8'>
+      <div className='sm:hidden flex justify-center mt-8'>
         <Button variant="link" className='text-primaryColor hover:scale-105' onClick={() => handleSeeMore(projects)}>Click to see more
           <ArrowUpRight className='w-4 h-4 inline-block' />
         </Button>
@@ -150,7 +107,7 @@ const Work = () => {
 
       {/* Popup Modal */}
       {selectedProject && (
-        <ProjectCard selectedProject={selectedProject} closeModal={closeModal} />
+        <ProjectModal selectedProject={selectedProject} closeModal={closeModal} />
       )}
     </section>
   );
